implement auth route; move functions to OOP AuthController (register, login, and refreshToken
