+++ api/controllers/authController.ts
@@ -1,14 +1,24 @@
 import { User } from '.prisma/client'
 import { Request, Response } from 'express'
-import jwt from 'jsonwebtoken'
+import jwt, { JwtPayload } from 'jsonwebtoken'
 import AuthService from '../services/authService'
 import UserService from '../services/userService'
-import { verifyPassword } from '../utils/hash'
-import arong2 from 'argon2'
+import argon2 from 'argon2'
 
-interface IPostUserBody {
+interface IUserLoginBody {
     email: string
     password: string
+}
+
+interface IUserRegistrationBody {
+    email: string
+    last: string
+    first: string
+    password: string
+}
+
+interface IUserRefreshTokenBody {
+    refreshToken: string
 }
 
 export default class AuthController {
@@ -33,16 +43,16 @@ export default class AuthController {
         return jwt.verify(token, process.env.JWT_SECRET as string)
     }
 
-    async hashToken(token: string): Promise<string> {
+    static async hashToken(token: string): Promise<string> {
         return await argon2.hash(token)
     }
 
-    async verifyPassword(hash: string, password: string) {
+    static async verifyPassword(hash: string, password: string) {
         return await argon2.verify(hash, password)
     }
 
 
-    static async login(req: Request<any, any, IPostUserBody>, res: Response) {
+    static async login(req: Request<any, any, IUserLoginBody>, res: Response) {
         try {
             const { email, password } = req.body
 
@@ -53,7 +63,7 @@ export default class AuthController {
                 })
             }
 
-            const validatePassword = await verifyPassword(existingUser.password, password)
+            const validatePassword = await AuthController.verifyPassword(existingUser.password, password)
             if (!validatePassword) {
                 return res.status(401).json({
                     message: 'Invalid password'
@@ -62,6 +72,82 @@ export default class AuthController {
 
             const tokens = AuthController.generateTokens(existingUser)
             const whitelist = await AuthService.whitelistRefreshToken(existingUser, tokens.refreshToken)
+            return res.send({ tokens, whitelist }) // ? do we need to send whitelist?
+        } catch (error) {
+            if (error instanceof Error) {
+                res.status(500).send({
+                    message: 'something went wrong'
+                })
+                throw error
+            }
+        }
+    }
+
+    static async register(req: Request<any, any, IUserRegistrationBody>, res: Response) {
+        try {
+            const { email, last, first, password } = req.body
+            if (!email || !password) {
+                return res.status(400).send({
+                    message: 'email and password are required'
+                })
+            }
+
+            const existingUser = await UserService.findUserByEmail(email)
+            if (existingUser) {
+                return res.status(400).send({
+                    message: 'user already exists'
+                })
+            }
+
+            const user = await UserService.createUser(last, first, email, password)
+            const tokens = AuthController.generateTokens(user)
+            const whitelist = await AuthService.whitelistRefreshToken(user, tokens.refreshToken)
+            res.send({ tokens, whitelist }) // ? do we need to send the whitelist?
+        } catch (error) {
+            if (error instanceof Error) {
+                res.status(500).send({
+                    message: 'something went wrong'
+                })
+                throw error
+            }
+        }
+    }
+
+    static async refreshToken(req: Request<any, any, IUserRefreshTokenBody>, res: Response) {
+        try {
+            const { refreshToken } = req.body
+            if (!refreshToken) {
+                return res.status(400).send({
+                    message: 'refreshToken is required'
+                })
+            }
+            // todo this will probably not work due to the current setup for findrefreshToken
+            // todo may need uuidv4 for jti
+            const payload = jwt.verify(refreshToken, process.env.refresh_token_secret as string) as JwtPayload
+            const savedRefreshToken = await AuthService.findRefreshToken(payload.jti as string)
+            if (!savedRefreshToken || savedRefreshToken.revoked) {
+                return res.status(401).send({
+                    message: 'unauthorized'
+                })
+            }
+
+            const hashedToken = await AuthController.hashToken(refreshToken)
+            if (hashedToken !== savedRefreshToken.hashedToken) {
+                return res.status(401).send({
+                    message: 'unauthorized'
+                })
+            }
+
+            const user = await UserService.findUserById(payload.id)
+            if (!user) {
+                return res.status(401).send({
+                    message: 'unauthorized'
+                })
+            }
+
+            const deleteRefreshTokens = await AuthService.deleteRefreshToken(payload.jti as string)
+            const tokens = AuthController.generateTokens(user)
+            const whitelist = await AuthService.whitelistRefreshToken(user, tokens.refreshToken)
             return res.send({ tokens, whitelist }) // ? do we need to send whitelist?
         } catch (error) {
             if (error instanceof Error) {
